Extract `InlineCode` component into own file
