+++ website/src/templates/models.js
@@ -5,7 +5,7 @@ import Title from '../components/title'
 import Section from '../components/section'
 import Button from '../components/button'
 import Aside from '../components/aside'
-import { InlineCode } from '../components/code'
+import { InlineCode } from '../components/inlineCode'
 import CodeBlock from '../components/codeBlock'
 import { Table, Tr, Td, Th } from '../components/table'
 import Tag from '../components/tag'
