+++ website/src/templates/universe.js
@@ -8,7 +8,7 @@ import Grid from '../components/grid'
 import Button from '../components/button'
 import Icon from '../components/icon'
 import Tag from '../components/tag'
-import { InlineCode } from '../components/code'
+import { InlineCode } from '../components/inlineCode'
 import CodeBlock from '../components/codeBlock'
 import Aside from '../components/aside'
 import Sidebar from '../components/sidebar'
